started working on the test

+++ src/test/java/africa/nkwadoma/nkwadoma/infrastructure/adapters/output/loanManagement/LoanAdapterTest.java
@@ -8,6 +8,7 @@ import africa.nkwadoma.nkwadoma.domain.model.loan.Loanee;
 import africa.nkwadoma.nkwadoma.infrastructure.adapters.output.persistence.entity.loanEntity.LoaneeEntity;
 import africa.nkwadoma.nkwadoma.infrastructure.adapters.output.persistence.repository.loan.LoaneeRepository;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -34,24 +35,27 @@ class LoanAdapterTest {
         LoaneeEntity loaneeEntity = new LoaneeEntity();
         loaneeEntity = loaneeRepository.save(loaneeEntity);
         UserIdentity userIdentity = new UserIdentity();
+        Loanee loanee = new Loanee();
+        loanee.setUserIdentity(userIdentity);
 
         loan = new Loan();
         loan.setLoanAccountId("account id");
         loan.setLoanee(Loanee.builder().id(loaneeEntity.getId()).build());
         loan.setStartDate(LocalDateTime.now());
+        loan.setLoanee(loanee);
 
     }
     @Test
     void saveLoan() throws MeedlException {
         Loan savedLoan = null;
-//        try {
+        try {
             savedLoan = loanOutputPort.save(loan);
             savedLoanId = savedLoan.getId();
             log.info("Saved loan: {} ", savedLoan.getId());
-//        } catch (MeedlException e) {
-//            log.error("Error saving loan {}", e.getMessage());
-//            throw new RuntimeException(e.getMessage());
-//        }
+        } catch (MeedlException e) {
+            log.error("Error saving loan {}", e.getMessage());
+            throw new RuntimeException(e.getMessage());
+        }
 
         assertNotNull(savedLoan);
         assertNotNull(savedLoan.getId());
@@ -72,7 +76,11 @@ class LoanAdapterTest {
     }
     @AfterEach
     void tearDown() {
-        loanOutputPort.deleteById(savedLoanId);
-        loaneeRepository.deleteById(loan.getLoanee().getId());
+        if (StringUtils.isNotEmpty(savedLoanId)) {
+            loanOutputPort.deleteById(savedLoanId);
+        }
+        if (StringUtils.isNotEmpty(loan.getLoanee().getId())) {
+            loaneeRepository.deleteById(loan.getLoanee().getId());
+        }
     }
 }
